+++ py/jupyterlite-pyodide-lock/src/jupyterlite_pyodide_lock/constants.py
