+++ py/jupyterlite-pyodide-lock/src/jupyterlite_pyodide_lock/lockers/browser.py
