+++ py/jupyterlite-pyodide-lock/src/jupyterlite_pyodide_lock/lockers/handlers/lock.html.j2
